Factor out ActionExecutionClient (#5796)
