+++ evaluation/benchmarks/biocoder/run_infer.py
@@ -61,7 +61,7 @@ def get_config(
     config = AppConfig(
         default_agent=metadata.agent_class,
         run_as_openhands=False,
-        runtime='eventstream',
+        runtime='docker',
         max_iterations=metadata.max_iterations,
         sandbox=SandboxConfig(
             base_container_image=BIOCODER_BENCH_CONTAINER_IMAGE,
