+++ evaluation/benchmarks/ml_bench/run_infer.py
@@ -80,7 +80,7 @@ def get_config(
     config = AppConfig(
         default_agent=metadata.agent_class,
         run_as_openhands=False,
-        runtime='eventstream',
+        runtime='docker',
         max_iterations=metadata.max_iterations,
         sandbox=SandboxConfig(
             base_container_image='public.ecr.aws/i5g0m1f6/ml-bench',
