+++ openhands/runtime/impl/modal/modal_runtime.py
@@ -1,8 +1,7 @@
 import os
 import tempfile
-import threading
 from pathlib import Path
-from typing import Callable, Generator
+from typing import Callable
 
 import modal
 import requests
@@ -10,9 +9,8 @@ import tenacity
 
 from openhands.core.config import AppConfig
 from openhands.events import EventStream
-from openhands.runtime.impl.eventstream.eventstream_runtime import (
-    EventStreamRuntime,
-    LogStreamer,
+from openhands.runtime.impl.action_execution.action_execution_client import (
+    ActionExecutionClient,
 )
 from openhands.runtime.plugins import PluginRequirement
 from openhands.runtime.utils.command import get_remote_startup_command
@@ -21,52 +19,13 @@ from openhands.runtime.utils.runtime_build import (
     prep_build_folder,
 )
 from openhands.utils.async_utils import call_sync_from_async
+from openhands.utils.tenacity_stop import stop_if_should_exit
 
 # FIXME: this will not work in HA mode. We need a better way to track IDs
 MODAL_RUNTIME_IDS: dict[str, str] = {}
 
 
-# Modal's log generator returns strings, but the upstream LogBuffer expects bytes.
-def bytes_shim(string_generator) -> Generator[bytes, None, None]:
-    for line in string_generator:
-        yield line.encode('utf-8')
-
-
-class ModalLogStreamer(LogStreamer):
-    """Streams Modal sandbox logs to stdout.
-
-    This class provides a way to stream logs from a Modal sandbox directly to stdout
-    through the provided logging function.
-    """
-
-    def __init__(
-        self,
-        sandbox: modal.Sandbox,
-        logFn: Callable,
-    ):
-        self.log = logFn
-        self._stop_event = threading.Event()
-        self.log_generator = bytes_shim(sandbox.stderr)
-
-        # Start the stdout streaming thread
-        self.stdout_thread = threading.Thread(target=self._stream_logs)
-        self.stdout_thread.daemon = True
-        self.stdout_thread.start()
-
-    def _stream_logs(self):
-        """Stream logs from the Modal sandbox."""
-        try:
-            for log_line in self.log_generator:
-                if self._stop_event.is_set():
-                    break
-                if log_line:
-                    decoded_line = log_line.decode('utf-8').rstrip()
-                    self.log('debug', f'[inside sandbox] {decoded_line}')
-        except Exception as e:
-            self.log('error', f'Error streaming modal logs: {e}')
-
-
-class ModalRuntime(EventStreamRuntime):
+class ModalRuntime(ActionExecutionClient):
     """This runtime will subscribe the event stream.
 
     When receive an event, it will send the event to runtime-client which run inside the Modal sandbox environment.
@@ -116,14 +75,9 @@ class ModalRuntime(EventStreamRuntime):
         # This value is arbitrary as it's private to the container
         self.container_port = 3000
 
-        self.session = requests.Session()
         self.status_callback = status_callback
         self.base_container_image_id = self.config.sandbox.base_container_image
         self.runtime_container_image_id = self.config.sandbox.runtime_container_image
-        self.action_semaphore = threading.Semaphore(1)  # Ensure one action at a time
-
-        # Buffer for container logs
-        self.log_streamer: LogStreamer | None = None
 
         if self.config.sandbox.runtime_extra_deps:
             self.log(
@@ -131,7 +85,7 @@ class ModalRuntime(EventStreamRuntime):
                 f'Installing extra user-provided dependencies in the runtime image: {self.config.sandbox.runtime_extra_deps}',
             )
 
-        self.init_base_runtime(
+        super().__init__(
             config,
             event_stream,
             sid,
@@ -170,7 +124,6 @@ class ModalRuntime(EventStreamRuntime):
 
             self.send_status_message('STATUS$CONTAINER_STARTED')
 
-        self.log_streamer = ModalLogStreamer(self.sandbox, self.log)
         if self.sandbox is None:
             raise Exception('Sandbox not initialized')
         tunnel = self.sandbox.tunnels()[self.container_port]
@@ -186,6 +139,20 @@ class ModalRuntime(EventStreamRuntime):
 
         if not self.attach_to_existing:
             self.send_status_message(' ')
+
+    def _get_action_execution_server_host(self):
+        return self.api_url
+
+    @tenacity.retry(
+        stop=tenacity.stop_after_delay(120) | stop_if_should_exit(),
+        retry=tenacity.retry_if_exception_type(
+            (ConnectionError, requests.exceptions.ConnectionError)
+        ),
+        reraise=True,
+        wait=tenacity.wait_fixed(2),
+    )
+    def _wait_until_alive(self):
+        self.check_if_alive()
 
     def _get_image_definition(
         self,
@@ -292,11 +259,7 @@ echo 'export INPUTRC=/etc/inputrc' >> /etc/bash.bashrc
 
     def close(self):
         """Closes the ModalRuntime and associated objects."""
-        if self.log_streamer:
-            self.log_streamer.close()
-
-        if self.session:
-            self.session.close()
+        super().close()
 
         if not self.attach_to_existing and self.sandbox:
             self.sandbox.terminate()
