+++ openhands/runtime/impl/runloop/runloop_runtime.py
@@ -1,82 +1,26 @@
 import logging
-import threading
-import time
 from typing import Callable
 
-import requests
 import tenacity
 from runloop_api_client import Runloop
 from runloop_api_client.types import DevboxView
 from runloop_api_client.types.shared_params import LaunchParameters
 
 from openhands.core.config import AppConfig
-from openhands.core.exceptions import (
-    AgentRuntimeNotReadyError,
-    AgentRuntimeUnavailableError,
-)
 from openhands.core.logger import openhands_logger as logger
 from openhands.events import EventStream
-from openhands.runtime.impl.eventstream.eventstream_runtime import EventStreamRuntime
+from openhands.runtime.impl.action_execution.action_execution_client import (
+    ActionExecutionClient,
+)
 from openhands.runtime.plugins import PluginRequirement
 from openhands.runtime.utils.command import get_remote_startup_command
-from openhands.runtime.utils.log_streamer import LogStreamer
-from openhands.runtime.utils.request import send_request
 from openhands.utils.tenacity_stop import stop_if_should_exit
 
 CONTAINER_NAME_PREFIX = 'openhands-runtime-'
 
 
-class RunloopLogStreamer(LogStreamer):
-    """Streams Runloop devbox logs to stdout.
-
-    This class provides a way to stream logs from a Runloop devbox directly to stdout
-    through the provided logging function.
-    """
-
-    def __init__(
-        self,
-        runloop_api_client: Runloop,
-        devbox_id: str,
-        logFn: Callable,
-    ):
-        self.runloop_api_client = runloop_api_client
-        self.devbox_id = devbox_id
-        self.log = logFn
-        self.log_index = 0
-        self._stop_event = threading.Event()
-
-        # Start the stdout streaming thread
-        self.stdout_thread = threading.Thread(target=self._stream_logs)
-        self.stdout_thread.daemon = True
-        self.stdout_thread.start()
-
-    def _stream_logs(self):
-        """Stream logs from the Runloop devbox."""
-        try:
-            while True:
-                raw_logs = self.runloop_api_client.devboxes.logs.list(
-                    self.devbox_id
-                ).logs[self.log_index :]
-                logs = [
-                    log.message
-                    for log in raw_logs
-                    if log.message and log.cmd_id is None
-                ]
-
-                self.log_index += len(raw_logs)
-                if self._stop_event.is_set():
-                    break
-                if logs:
-                    for log_line in logs:
-                        self.log('debug', f'[inside devbox] {log_line}')
-
-                time.sleep(1)
-        except Exception as e:
-            self.log('error', f'Error streaming runloop logs: {e}')
-
-
-class RunloopRuntime(EventStreamRuntime):
-    """The RunloopRuntime class is an EventStreamRuntime that utilizes Runloop Devbox as a runtime environment."""
+class RunloopRuntime(ActionExecutionClient):
+    """The RunloopRuntime class is an DockerRuntime that utilizes Runloop Devbox as a runtime environment."""
 
     _sandbox_port: int = 4444
     _vscode_port: int = 4445
@@ -98,10 +42,8 @@ class RunloopRuntime(EventStreamRuntime):
         self.runloop_api_client = Runloop(
             bearer_token=config.runloop_api_key,
         )
-        self.session = requests.Session()
         self.container_name = CONTAINER_NAME_PREFIX + sid
-        self.action_semaphore = threading.Semaphore(1)  # Ensure one action at a time
-        self.init_base_runtime(
+        super().__init__(
             config,
             event_stream,
             sid,
@@ -112,8 +54,10 @@ class RunloopRuntime(EventStreamRuntime):
             headless_mode,
         )
         # Buffer for container logs
-        self.log_streamer: LogStreamer | None = None
         self._vscode_url: str | None = None
+
+    def _get_action_execution_server_host(self):
+        return self.api_url
 
     @tenacity.retry(
         stop=tenacity.stop_after_attempt(120),
@@ -203,15 +147,11 @@ class RunloopRuntime(EventStreamRuntime):
             port=self._sandbox_port,
         )
 
-        # Hook up logs
-        self.log_streamer = RunloopLogStreamer(
-            self.runloop_api_client, self.devbox.id, logger.info
-        )
         self.api_url = tunnel.url
         logger.info(f'Container started. Server url: {self.api_url}')
 
         # End Runloop connect
-        # NOTE: Copied from EventStreamRuntime
+        # NOTE: Copied from DockerRuntime
         logger.info('Waiting for client to become ready...')
         self.send_status_message('STATUS$WAITING_FOR_CLIENT')
         self._wait_until_alive()
@@ -230,27 +170,10 @@ class RunloopRuntime(EventStreamRuntime):
         reraise=(ConnectionRefusedError,),
     )
     def _wait_until_alive(self):
-        if not self.log_streamer:
-            raise AgentRuntimeNotReadyError('Runtime client is not ready.')
-        response = send_request(
-            self.session,
-            'GET',
-            f'{self.api_url}/alive',
-            timeout=5,
-        )
-        if response.status_code == 200:
-            return
-        else:
-            msg = f'Action execution API is not alive. Response: {response}'
-            logger.error(msg)
-            raise AgentRuntimeUnavailableError(msg)
+        super().check_if_alive()
 
     def close(self, rm_all_containers: bool | None = True):
-        if self.log_streamer:
-            self.log_streamer.close()
-
-        if self.session:
-            self.session.close()
+        super().close()
 
         if self.attach_to_existing:
             return
@@ -260,42 +183,24 @@ class RunloopRuntime(EventStreamRuntime):
 
     @property
     def vscode_url(self) -> str | None:
-        if self.vscode_enabled and self.devbox and self.devbox.status == 'running':
-            if self._vscode_url is not None:
-                return self._vscode_url
-
-            try:
-                with send_request(
-                    self.session,
-                    'GET',
-                    f'{self.api_url}/vscode/connection_token',
-                    timeout=10,
-                ) as response:
-                    response_json = response.json()
-                    assert isinstance(response_json, dict)
-                    if response_json['token'] is None:
-                        return None
-                    token = response_json['token']
-
-                self._vscode_url = (
-                    self.runloop_api_client.devboxes.create_tunnel(
-                        id=self.devbox.id,
-                        port=self._vscode_port,
-                    ).url
-                    + f'/?tkn={token}&folder={self.config.workspace_mount_path_in_sandbox}'
-                )
-
-                self.log(
-                    'debug',
-                    f'VSCode URL: {self._vscode_url}',
-                )
-
-                return self._vscode_url
-            except Exception as e:
-                self.log(
-                    'error',
-                    f'Failed to create vscode tunnel {e}',
-                )
-                return None
-        else:
+        if self._vscode_url is not None:  # cached value
+            return self._vscode_url
+        token = super().get_vscode_token()
+        if not token:
             return None
+        if not self.devbox:
+            return None
+        self._vscode_url = (
+            self.runloop_api_client.devboxes.create_tunnel(
+                id=self.devbox.id,
+                port=self._vscode_port,
+            ).url
+            + f'/?tkn={token}&folder={self.config.workspace_mount_path_in_sandbox}'
+        )
+
+        self.log(
+            'debug',
+            f'VSCode URL: {self._vscode_url}',
+        )
+
+        return self._vscode_url
